+++ ChangeLog.md
@@ -4,6 +4,7 @@ This document is created to record all changes.
 -------------------
 ## [Unrelease]
 ### Added
+- Add the header file `includes/ObserverPattern.h` to define the class `myAbstractObserver` for observer pattern.
 
 ### Changed